check dc group for new additional column on orders

+++ src/main/java/com/turkcell/bootcamp/e_commerce/entities/Payment.java
@@ -7,7 +7,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.math.BigDecimal;
-import java.util.Date;
+import java.sql.Date;
 
 @Table(name="payments")
 @Entity
@@ -20,7 +20,7 @@ public class Payment {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
  @Column(name="amount")
-    private BigDecimal amount;//bak
+    private double amount;//bak
 
  @Column(name="date")
     private Date date;//bak
